+++ module3/presentation_hazards.md
@@ -134,6 +134,36 @@ int main() {
 ```
 
 ___
+<!-- .slide: style="font-size: 0.85em" -->
+
+## Zadanie
+
+Napisz funkcję `foo()`. Ma ona przyjmować shared_ptr na int i ma przypisać wartość 20 do wskazywanego przez niego obiektu.
+
+Ponadto `foo()` ma wyświetlić wartość inta wskazywanego przez wskaźnik oraz liczbę shared_ptrów, które wskazują na ten obiekt.
+
+Wyświetl także to samo w `main()` przed i po zawołaniu `foo()`.
+
+```cpp
+#include <iostream>
+#include <memory>
+
+// TODO: Implement foo()
+// It should take shared_ptr to int and assign value 20 to the pointed int.
+// It should also display the value of this int and the number of how many pointers are pointing to it - use `use_count()`.
+// Display the same information in main() before and after calling foo()
+
+int main() {
+    std::shared_ptr<int> number = std::make_shared<int>(10);
+    // display the value under number pointer and use_count() of it
+    foo(number);
+    // display the value under number pointer and use_count() of it
+
+    return 0;
+}
+```
+
+___
 
 ## Inteligentne wskaźniki rozwiązaniem wszystkich problemów?
