+++ module3/task3.cpp
@@ -1,0 +1,16 @@
+#include <iostream>
+#include <memory>
+
+// TODO: Implement foo()
+// It should take shared_ptr to int and assign value 20 to the pointed int.
+// It should also display the value of this int and the number of how many pointers are pointing to it - use `use_count()`.
+// Display the same information in main() before and after calling foo()
+
+int main() {
+    std::shared_ptr<int> number = std::make_shared<int>(10);
+    // display the value under number pointer and use_count() of it
+    foo(number);
+    // display the value under number pointer and use_count() of it
+
+    return 0;
+}
